+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Fixed not all architectures being published to the Flatpak repo
 - Fixed MMC1 games (e.g. The Legend of Zelda) freezing after rewinding
 - Fixed controller focus getting stuck in the file picker's search bar
+- Fixed Escape quitting to the main menu instead of opening the in-game menu after resuming the game with Enter
 
 ## [0.17.0] - 2026-08-23